requirements.txt
anywidget==0.9.7
hdbscan>=0.8.33
ipywidgets==8.1.2
langchain>=0.1.5
nltk>=3.8.1
numpy>=1.23.5
openai>=1.23.1
pandas>=2.2.0
pathos>=0.3.0
protobuf>=4.25.2
PyYAML>=6.0
tiktoken>=0.7.0
tqdm>=4.66.1
traitlets>=5.14.1
umap-learn>=0.5.5
jupyterlab_widgets>=3.0.10
yaspin>=3.0.2
requests>=2.28.2
jupyterlab>=4.2
python-dotenv>=1.0.1
